+++ data/src/main/java/com/roacult/kero/oxxy/projet2eme/network/entities/UpdateUser.kt
@@ -1,4 +1,4 @@
 package com.roacult.kero.oxxy.projet2eme.network.entities
 
-data class UpdateUserParam(val id:Long ,val  fname:String ,val lname:String , val picture:String , val ispublic:Boolean, val password:String?  )
+data class UpdateUserParam(val id:Int ,val  fname:String ,val lname:String , val picture:String , val ispublic:Boolean, val password:String?  )
 data class UpdateUserResult(val reponse :Int, val picture: String)
